Move ImportanceData to typing
