+++ recosamp/datasets/__init__.py
@@ -1,12 +1,11 @@
 from .debiased_triplet_dataset import DebiasedTripletDataset
-from .importance_filtered_triplet_dataset import ImportanceData, ImportanceFilteredTripletDataset
+from .importance_filtered_triplet_dataset import ImportanceFilteredTripletDataset
 from .triplet_dataset import TripletDataset
 from .uniform_triplet_dataset import UniformTripletDataset
 
 
 __all__ = [
     "DebiasedTripletDataset",
-    "ImportanceData",
     "ImportanceFilteredTripletDataset",
     "TripletDataset",
     "UniformTripletDataset",
